Replace lua with luajit in mpv
Fedora 22 ships lua 5.3, which is incompatible with lua 5.2. If only lua
5.3 is installed, mpv will not have OSD support or youtube-dl support.

In several issues in the issue tracker, the mpv development team shows
preference for keeping the code compatible with luajit instead of
migrating to lua 5.3.

Also, a future option may be to use compat-lua and compat-lua-devel,
which ships lua 5.1 in Fedora 22. But luajit seems nice.