Add an ffmpeg 2.8 spec file
ffmpeg 2.8 is still required by Firefox for H.264, MP3 and AAC support. Mimic
what ArchLinux does and create a specific ffmpeg2.8 package that is to be
installed in parallel with ffmpeg 3.0.

This package only contains the header files and shared libraries, installed to
specific subdirectories of /usr/include and /usr/lib64 so they won't collide
with ffmpeg 3.0.

By adding a file to /etc/ld.so.conf.d, some magic is done so the libraries in
/usr/lib64 take preference, but the ones in /usr/lib64/ffmpeg2.8 are also
included and can be found by Firefox when it tries to load them at runtime.